% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{bit_indexr}
\alias{bit_indexr}
\title{Binary indices}
\usage{
bit_indexr(n)
}
\arguments{
\item{n}{number of items to generate permutations from}
}
\value{
A matrix of logicals
}
\description{
Wraps around bit_indexr().
}
\keyword{internal}
