% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{center_circles}
\alias{center_circles}
\title{Center Circles}
\usage{
center_circles(pars)
}
\arguments{
\item{pars}{A matrix or data.frame of x coordinates, y coordinates, and
radii.}
}
\value{
A centered version of \code{pars}.
}
\description{
Center Circles
}
\keyword{internal}
