% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layout.R
\name{center_ellipses}
\alias{center_ellipses}
\title{Center Circles}
\usage{
center_ellipses(pars)
}
\arguments{
\item{pars}{A matrix or data.frame of x coordinates, y coordinates, minor
radius (a) and major radius (b).}
}
\value{
A centered version of \code{pars}.
}
\description{
Center Circles
}
\keyword{internal}
