% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layout.R
\name{center_layout}
\alias{center_layout}
\title{Center ellipses}
\usage{
center_layout(pars)
}
\arguments{
\item{pars}{a matrix or data.frame of x coordinates, y coordinates, minor
radius (a) and major radius (b)}
}
\value{
A centered version of \code{pars}.
}
\description{
Center ellipses
}
\keyword{internal}
