% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{diagError}
\alias{diagError}
\title{diagError}
\usage{
diagError(fit, orig, regionError = NULL)
}
\arguments{
\item{fit}{fitted values}

\item{orig}{original values}

\item{regionError}{regionError}
}
\value{
diagError
}
\description{
diagError
}
\keyword{internal}
