% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geometry.R
\name{dist_point_circle}
\alias{dist_point_circle}
\title{Distance from Point to Circle}
\usage{
dist_point_circle(x, y, h, k, a, b, phi)
}
\arguments{
\item{x}{X coordinate of point}

\item{y}{Y coordinate of point}

\item{h}{X coordinate of circle center}

\item{k}{Y coordinate of circle center}

\item{r}{Radius of circle center}
}
\description{
The function is vectorized.
}
\keyword{internal}
