% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{dont_plot}
\alias{dont_plot}
\title{Suppress plotting}
\usage{
dont_plot(x, ...)
}
\arguments{
\item{x}{object to call \code{\link[graphics:plot]{graphics::plot()}} on}

\item{...}{arguments to pass to \code{x}}
}
\value{
Invisibly returns whatever \code{plot(x)} would normally return, but
does not plot anything (which is the point).
}
\description{
Suppress plotting
}
\keyword{internal}
