% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{dont_print}
\alias{dont_print}
\title{Suppress printing}
\usage{
dont_print(x, ...)
}
\arguments{
\item{x}{object to (not) print}

\item{...}{arguments to \code{x}}
}
\value{
Invisibly returns the output of running print on \code{x}.
}
\description{
Suppress printing
}
\keyword{internal}
