% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{dummy_code}
\alias{dummy_code}
\title{Dummy code a data.frame}
\usage{
dummy_code(x, sep = "_", factor_names = TRUE)
}
\arguments{
\item{x}{a data.frame}

\item{sep}{character for separating dummy code factors and their levels
when constructing names}

\item{factor_names}{whether to include factor names when creating new
names for dummy codes}
}
\value{
A dummy-coded version of x.
}
\description{
Dummy code a data.frame
}
\keyword{internal}
