% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geometry.R
\name{ellipse}
\alias{ellipse}
\title{Plotting coordinates for an ellipse}
\usage{
ellipse(h, k, a, b = a, phi = 0, n = 200L)
}
\arguments{
\item{h}{x coordinates}

\item{k}{y coordinates}

\item{a}{semimajor axis}

\item{b}{semiminor axis}

\item{phi}{rotation}

\item{n}{number of plotting points}
}
\value{
A list of matrices of coordinates for the ellipses.
}
\description{
Plotting coordinates for an ellipse
}
\keyword{internal}
