% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geometry.R
\name{ellipse_bounding_box}
\alias{ellipse_bounding_box}
\title{Bounding box of a vector of rotated ellipses.}
\usage{
ellipse_bounding_box(h, k, a, b, phi)
}
\arguments{
\item{h, k}{ellipse center coordinates}

\item{a, b}{semi-major and semi-minor axes}

\item{phi}{rotation of the semi-major axis, in radians}
}
\description{
Bounding box of a vector of rotated ellipses.
}
\keyword{internal}
