% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main-old.R
\name{eulerr}
\alias{eulerr}
\title{Area-Proportional Euler Diagrams (defunct)}
\usage{
eulerr(...)
}
\arguments{
\item{\dots}{Ignored}
}
\description{
\emph{Note: This function is defunct; please use \code{\link{euler}}
instead.}
}
