% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eulerr_options.R
\name{eulerr_default_options}
\alias{eulerr_default_options}
\title{Default options for eulerr}
\usage{
eulerr_default_options()
}
\value{
default options for eulerr
}
\description{
Default options for eulerr
}
\keyword{internal}
