% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{eulerr_pal}
\alias{eulerr_pal}
\title{Set up a qualitative color palette}
\usage{
eulerr_pal(n)
}
\arguments{
\item{n}{number of colors to generate}
}
\value{
A string of hex colors.
}
\description{
Uses a custom color palette generated from \code{\link[qualpalr:qualpal]{qualpalr::qualpal()}},
which tries to provide distinct color palettes adapted to color vision
deficiency.
}
\keyword{internal}
