% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{fit_euler_diagram}
\alias{fit_euler_diagram}
\title{Fit an Euler diagram using the eunoia Rust library.}
\usage{
fit_euler_diagram(
  combo_names,
  combo_values,
  input,
  shape,
  loss,
  loss_aggregator,
  extraopt_threshold,
  tolerance,
  max_sets,
  seed
)
}
\description{
Fit an Euler diagram using the eunoia Rust library.
}
\keyword{internal}
