% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{fitted.eulerr}
\alias{fitted.eulerr}
\title{Fitted values from a eulerr fit}
\usage{
\method{fitted}{eulerr}(object, ...)
}
\arguments{
\item{object}{An object of class \code{eulerr}.}

\item{...}{Currently ignored}
}
\value{
A matrix of x and y coordinates and radiuses.
}
\description{
Returns a matrix of the circle centers in x and y coordinates, as well as
the radiuses.
}

