% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_constraints}
\alias{get_constraints}
\title{Set up constraints for optimization}
\usage{
get_constraints(newpars)
}
\arguments{
\item{newpars}{parameters from the first optimizer}
}
\value{
A list of lower and upper constraints
}
\description{
Set up constraints for optimization
}
\keyword{internal}
