% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.euler.R
\name{grobify}
\alias{grobify}
\title{Grobify Euler objects}
\usage{
grobify(x)
}
\arguments{
\item{x}{an \code{euler} object}
}
\value{
A \code{\link[grid:gList]{grid::gList()}} is returned.
}
\description{
Grobify Euler objects
}
\keyword{internal}
