% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_false}
\alias{is_false}
\title{Check If Object Is Strictly FALSE}
\usage{
is_false(x)
}
\arguments{
\item{x}{Object to check.}
}
\value{
Logical.
}
\description{
Check If Object Is Strictly FALSE
}
\keyword{internal}
