% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_real}
\alias{is_real}
\title{Check if vector is a real (numeric non-integer)}
\usage{
is_real(x, tol = .Machine$double.eps^0.5)
}
\arguments{
\item{x}{a vector}

\item{tol}{tolerance within which a value counts as a whole number}
}
\value{
A logical.
}
\description{
Check if vector is a real (numeric non-integer)
}
\keyword{internal}
