% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eulerr-deprecated.R
\name{label-deprecated}
\alias{label-deprecated}
\title{Compute Locations for Overlaps (deprecated)}
\usage{
label(x, labels)
}
\arguments{
\item{x}{An object of class 'euler'}

\item{labels}{An optional character vector of labels for the diagram.}
}
\value{
A numeric matrix of x and y coordinates for the labels, as well as
the quantities and proportions for the overlaps depicted in the labels.
}
\description{
Runs the same algorithm as in \code{\link[=plot.euler]{plot.euler()}} / \code{\link[=panel.euler.labels]{panel.euler.labels()}} to
label the overlaps in the Euler diagram. This is useful if you want to
use your own solution to plot the final diagram.
}
\details{
In the cases where \code{x} is a list of Euler diagrams (if the \code{by} argument
was used in the call to \code{\link[=euler]{euler()}}), this function returns a list of
matrices with coordinates.
}
\seealso{
\link{eulerr-deprecated}
}
\keyword{internal}
