% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.euler.R
\name{locate_centers}
\alias{locate_centers}
\title{Locate Centers of Circle Overlaps}
\usage{
locate_centers(x, y, a, b, phi, original.values, fitted.values)
}
\arguments{
\item{x}{X coordinates for the circle centers.}

\item{y}{Y coordinates for the circle centers.}

\item{phi}{Rotation of the ellipse (as the counter-clockwise angle from
the positive x-axis to the semi-major axis).}

\item{original.values}{Original values for the disjoint set combinations.}

\item{fitted.values}{Fitted values for the disjoint set combinations.}
}
\value{
A data frame with centers of the circle overlaps and their
respective original counts.
}
\description{
Locate Centers of Circle Overlaps
}
\keyword{internal}
