% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glyph-grobs.R
\name{makeContent.EulerSetLabels}
\alias{makeContent.EulerSetLabels}
\title{Re-place exterior set labels on the active device}
\usage{
\method{makeContent}{EulerSetLabels}(x)
}
\description{
Re-place exterior set labels on the active device
}
\keyword{internal}
