% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tag-grobs.R
\name{measure_tag_native}
\alias{measure_tag_native}
\title{Measure one tag's combined AABB in the current viewport's native units.}
\usage{
measure_tag_native(tag, padding_native)
}
\arguments{
\item{tag}{an \code{EulerTag} gTree}

\item{padding_native}{vertical separation between components, in native
units}
}
\description{
Measure one tag's combined AABB in the current viewport's native units.
}
\keyword{internal}
