% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{mix_colors}
\alias{mix_colors}
\title{Blend (average) colors}
\usage{
mix_colors(rcol_in)
}
\arguments{
\item{rcol_in}{a vector of R colors}
}
\value{
A single R color
}
\description{
Blend (average) colors
}
\keyword{internal}
