% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{n_sets}
\alias{n_sets}
\title{Get the number of sets in he input}
\usage{
n_sets(combinations)
}
\arguments{
\item{combinations}{A vector of combinations (see \code{\link[=euler]{euler()}}).}
}
\value{
The number of sets in the input
}
\description{
Get the number of sets in he input
}
\keyword{internal}
