% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{normalize_angle}
\alias{normalize_angle}
\title{Normalize an angle to [-pi, pi)}
\usage{
normalize_angle(x)
}
\arguments{
\item{x}{Angle in radians}
}
\value{
A normalized angle.
}
\description{
Normalize an angle to [-pi, pi)
}
\keyword{internal}
