% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{normalize_pars}
\alias{normalize_pars}
\title{Normalize parameters (semiaxes and rotation)}
\usage{
normalize_pars(m)
}
\arguments{
\item{m}{pars}
}
\value{
\code{m}, normalized
}
\description{
Normalize parameters (semiaxes and rotation)
}
\keyword{internal}
