% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{organisms}
\alias{organisms}
\title{Organisms}
\format{A \link{matrix} with 7 observations, consisting of various organisms,
and 5 variables: \emph{animal, mammal, plant, sea}, and, \emph{spiny},
indicating whether the organism belongs to the category or not.}
\source{
\url{https://github.com/sysbio-bioinf/VennMaster/blob/master/data_examples/deploy/example1.list}
}
\usage{
organisms
}
\description{
Example data from the \strong{VennMaster} package.
}
\details{
Note that this data is difficult to fit using an Euler diagram, even
if we use ellipses, which is clear if one chooses to study the various
overlaps in the resulting diagrams.
}
\keyword{datasets}
