% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pain}
\alias{pain}
\title{Pain distribution data}
\format{A flat \link{table} (cross-table) with with sex in columns and
pain distribution in rows and integer counts making up the cells of the
table.}
\source{
Westergren H, Larsson J, Freeman M, Carlsson A, Jöud A,
Malmström E-M. Sex-based differences in pain distribution in a cohort of
patients with persistent post-traumatic neck pain. Disability and
Rehabilitation. 2017 Jan 27; Available
from: \url{http://www.tandfonline.com/eprint/RK8tusJswJCsa9YakGDW/full}
}
\usage{
pain
}
\description{
Data from a study on pain distribution for patients with
persistent neck pain in relation to a whiplash trauma.
}
\section{Disclaimer}{

Note that the maintainer of this package is an author of the source for
this data.
}

\keyword{datasets}
