% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.euler.R
\name{panel.euler.circles}
\alias{panel.euler.circles}
\title{Panel Function for Circles}
\usage{
panel.euler.circles(x, y, r, border = "black", fill = "transparent", ...,
  col, font, fontface)
}
\arguments{
\item{x}{X coordinates for the circle centers.}

\item{y}{Y coordinates for the circle centers.}

\item{r}{Radii.}

\item{border}{Border color.}

\item{fill}{Circle fill.}

\item{...}{Passed on to \code{\link[grid:grid.circle]{grid::grid.circle()}}.}

\item{col}{Ignored}

\item{font}{Ignored}

\item{fontface}{Ignored}
}
\value{
Plots circles inside a trellis panel.
}
\description{
Panel Function for Circles
}
\seealso{
\code{\link[grid:grid.circle]{grid::grid.circle()}}.
}
