% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eulerr_deprecated.R
\name{panel.euler.circles_deprecated}
\alias{panel.euler.circles_deprecated}
\title{Panel Function for Euler Circles (deprecated)}
\usage{
panel.euler.circles(x, y, r, border, fill, fill_alpha, ...,
  identifier, name.type, col, font, fontface)
}
\arguments{
\item{x}{X coordinates for the centers.}

\item{y}{Y coordinates for the centers.}

\item{r}{Radius of the circle}

\item{border}{Border color.}

\item{fill}{Circle fill.}

\item{fill_alpha}{Fill opacity}

\item{...}{Passed on to \code{\link[grid:grid.circle]{grid::grid.circle()}}.}

\item{col}{Ignored}

\item{font}{Ignored}

\item{fontface}{Ignored}

\item{identifier}{A character string that is prepended to the name of the
grob that is created.}

\item{name.type}{A character value indicating whether the name of the grob
should have panel or strip information added to it. Typically either
\code{"panel"}, \code{"strip"}, \code{"strip.left"}, or \code{""} (for no extra information).}
}
\value{
Plots circles inside a trellis panel.
}
\description{
Panel Function for Euler Circles (deprecated)
}
\seealso{
\code{\link[grid:grid.circle]{grid::grid.circle()}}.

\link{eulerr_deprecated}
}
\keyword{internal}
