% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eulerr_deprecated.R
\name{panel.euler.ellipses_deprecated}
\alias{panel.euler.ellipses_deprecated}
\title{Panel Function for Euler Ellipses (deprecated)}
\usage{
panel.euler.ellipses(x, y, ra, rb, phi, fill = "transparent",
  fill_alpha = 0.4, border = "black", mode = c("split", "overlay"),
  identifier = NULL, n = 200, fitted.values, ..., name.type = "panel",
  col, font, fontface)
}
\arguments{
\item{x}{X coordinates for the centers.}

\item{y}{Y coordinates for the centers.}

\item{ra}{Semi-major axes.}

\item{rb}{Semi-minor axes.}

\item{phi}{Rotation of the ellipse (as the counter-clockwise angle from
the positive x-axis to the semi-major axis).}

\item{border}{Border color.}

\item{fill}{Ellipse fill.}

\item{n}{Number of vertices to draw for each ellipse.}

\item{...}{Passed on to \code{\link[grid:grid.polygon]{grid::grid.polygon()}}.}

\item{col}{Ignored}

\item{font}{Ignored}

\item{fontface}{Ignored}

\item{identifier}{A character string that is prepended to the name of the
grob that is created.}

\item{name.type}{A character value indicating whether the name of the grob
should have panel or strip information added to it. Typically either
\code{"panel"}, \code{"strip"}, \code{"strip.left"}, or \code{""} (for no extra information).}

\item{mode}{"\code{split}", the default, splits up the diagram into individual
polygons and blends the colors of the overlapping shapes using
color averaging in the CIELAB color space. "\code{overlay}" superposes
sets and should be used in conjunction with a suitable \code{fill_alpha} value.}

\item{original.values}{Original values for the disjoint set combinations.}

\item{fitted.values}{Fitted values for the disjoint set combinations.}
}
\value{
Plots ellipses inside a trellis panel.
}
\description{
Panel Function for Euler Ellipses (deprecated)
}
\seealso{
\code{\link[grid:grid.polygon]{grid::grid.polygon()}}.

\link{eulerr_deprecated}
}
\keyword{internal}
