% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.euler.R
\name{panel.euler.labels}
\alias{panel.euler.labels}
\title{Panel function for Euler diagram labels}
\usage{
panel.euler.labels(x, y, ra, rb, phi, labels, quantities = FALSE,
  original.values, fitted.values, ...)
}
\arguments{
\item{x}{X coordinates for the centers.}

\item{y}{Y coordinates for the centers.}

\item{ra}{Semi-major axes.}

\item{rb}{Semi-minor axes.}

\item{phi}{Rotation of the ellipse (as the counter-clockwise angle from
the positive x-axis to the semi-major axis).}

\item{labels}{Labels.}

\item{quantities}{Plots the original values for the disjoint set combinations
(\code{original.values}). Can also be a list, in which the contents of the list
will be passed on to \code{\link[lattice:panel.text]{lattice::panel.text()}} to modify the appearance of
the quantity labels.}

\item{original.values}{Original values for the disjoint set combinations.}

\item{fitted.values}{Fitted values for the disjoint set combinations.}

\item{...}{Arguments passed on to \code{\link[=panel.text]{panel.text()}}}
}
\value{
Computes and plots labels or quantities inside the centers of the
ellipses' overlaps.
}
\description{
Panel function for Euler diagram labels
}
