% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eulerr_deprecated.R
\name{panel.euler.labels_deprecated}
\alias{panel.euler.labels_deprecated}
\title{Panel Function for Euler Diagram Labels (deprecated)}
\usage{
panel.euler.labels(x, y, ra, rb, phi, labels, quantities, original.values,
  fitted.values, ...)
}
\arguments{
\item{...}{Arguments passed on to \code{\link[=panel.text]{panel.text()}}}
}
\value{
Computes and plots labels or quantities inside the centers of the
ellipses' overlaps.
}
\description{
Panel Function for Euler Diagram Labels (deprecated)
}
\seealso{
\link{eulerr_deprecated}
}
\keyword{internal}
