% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{plants}
\alias{plants}
\title{Plants}
\format{A list with 33,721 plants, each containing a character vector listing
the states in the US and Canada in which they occur. The names in the
list specify the species or genus of the plant.}
\source{
USDA, NRCS. 2008. The PLANTS Database (\url{http://plants.usda.gov/}, 31
December 2008). National Plant Data Center, Baton Rouge, LA 70874-4490 USA.

Dua, D. and Karra Taniskidou, E. (2017). UCI Machine Learning
Repository \url{http://archive.ics.uci.edu/ml}. Irvine, CA: University of
California, School of Information and Computer Science.
}
\usage{
plants
}
\description{
Data on plants and the states in the US and Canada they occur in.
}
\keyword{datasets}
