% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geometry.R
\name{poly_clip}
\alias{poly_clip}
\title{Polygon Clipping}
\usage{
poly_clip(a, b, op = c("intersection", "union", "minus", "xor"))
}
\arguments{
\item{a}{polygon}

\item{b}{polygon}

\item{op}{operation}
}
\value{
A list of lists.
}
\description{
This function is provided to efficiently and safely handle clipping
operations. It wraps around \code{\link[polyclip:polyclip]{polyclip::polyclip()}}, which is an
interface to the \strong{Clipper} C++ library.
}
\keyword{internal}
