% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.euler.R
\name{prepanel.euler}
\alias{prepanel.euler}
\title{Prepanel Function for Euler Diagrams}
\usage{
prepanel.euler(x, y, r = NULL, ra = NULL, rb = NULL, phi = NULL,
  subscripts, ...)
}
\arguments{
\item{x}{X coordinates for the circle centers.}

\item{y}{Y coordinates for the circle centers.}

\item{r}{Radii.}

\item{ra}{Semi-major axes.}

\item{rb}{Semi-minor axes.}

\item{phi}{Rotation of the ellipse (as the counter-clockwise angle from
the positive x-axis to the semi-major axis).}

\item{subscripts}{A vector of subscripts (See \code{\link[lattice:xyplot]{lattice::xyplot()}}).}

\item{...}{Ignored.}
}
\value{
A list of \code{xlim} and \code{ylim} items.
}
\description{
Prepanel Function for Euler Diagrams
}
