% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eulerr_deprecated.R
\name{prepanel.euler_deprecated}
\alias{prepanel.euler_deprecated}
\title{Prepanel Function for Euler Diagrams (deprecated)}
\usage{
prepanel.euler(x, y, ra, rb, phi, subscripts, ...)
}
\arguments{
\item{x}{X coordinates for the centers.}

\item{y}{Y coordinates for the centers.}

\item{ra}{Semi-major axes.}

\item{rb}{Semi-minor axes.}

\item{phi}{Rotation of the ellipse (as the counter-clockwise angle from
the positive x-axis to the semi-major axis).}
}
\value{
A list of \code{xlim} and \code{ylim} items.
}
\description{
Prepanel Function for Euler Diagrams (deprecated)
}
\seealso{
\link{eulerr_deprecated}
}
\keyword{internal}
