% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{print.eulerr}
\alias{print.eulerr}
\title{Print eulerr fits}
\usage{
\method{print}{eulerr}(x, round = 3, ...)
}
\arguments{
\item{x}{Euler diagram specification from \pkg{eulerr}.}

\item{round}{Number of decimal places to round to.}

\item{...}{Arguments passed to \code{\link[base]{print.data.frame}}.}
}
\value{
Prints the results of the fit.
}
\description{
Prints a data frame of the original set relationships and the fitted
values as well as diagError and stress statistics.
}

