% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.euler.R
\name{print_diagram}
\alias{print_diagram}
\title{Print (plot) Euler diagram}
\usage{
print_diagram(x)
}
\arguments{
\item{x}{an object of class \code{'euler'}, usually the output of
\code{\link[=plot.euler]{plot.euler()}}.}
}
\value{
A plot is drawn on the current device using \code{\link[grid:Grid]{grid::Grid()}}.
}
\description{
This function is responsible for the actual plotting of
\code{'euler_diagram'} objects created through \code{\link[=plot.euler]{plot.euler()}}.
}
