% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{qualpalr_pal}
\alias{qualpalr_pal}
\title{Set up a qualitative color palette}
\usage{
qualpalr_pal(n)
}
\arguments{
\item{n}{Number of Colors to Generate}
}
\value{
A string of hex colors
}
\description{
Uses a custom color palette generated from \code{\link[qualpalr:qualpal]{qualpalr::qualpal()}},
which tries to provide distinct color palettes adapted to color vision
deficiency.
}
\keyword{internal}
