% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{replace_list}
\alias{replace_list}
\title{Replace (refresh) a list}
\usage{
replace_list(old, new)
}
\arguments{
\item{old}{the original list}

\item{new}{the things to update \code{old} with}
}
\value{
A refreshed list.
}
\description{
Unlike \code{update_list}, this function only modifies, and does not add,
items in the list.
}
\keyword{internal}
