% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rescale}
\alias{rescale}
\title{Rescale values to new range}
\usage{
rescale(x, new_min, new_max)
}
\arguments{
\item{x}{Numeric vector}

\item{new_min}{New min}

\item{new_max}{New max}
}
\value{
Rescaled vector
}
\description{
Rescale values to new range
}
\keyword{internal}
