% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/euler-methods.R
\name{residuals.euler}
\alias{residuals.euler}
\title{Residuals of euler object}
\usage{
\method{residuals}{euler}(object, dense = FALSE, ...)
}
\arguments{
\item{object}{object of class \code{'euler'}}

\item{dense}{same meaning as in \code{\link[=fitted.euler]{fitted.euler()}}}

\item{...}{ignored}
}
\value{
A named numeric vector of residuals (input minus fit) keyed by
combination label.
}
\description{
Residuals of euler object
}
