% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eulerr_methods.R
\name{residuals.eulerr}
\alias{residuals.eulerr}
\title{Squared residuals from a eulerr fit.}
\usage{
\method{residuals}{eulerr}(object, ...)
}
\arguments{
\item{object}{A euler specification of class \code{eulerr}.}

\item{\dots}{Currently ignored.}
}
\value{
Squared residuals.
}
\description{
Squared residuals from a eulerr fit.
}
\examples{
fit <- eulerr(c("A" = 1, "B" = 0.4, "C" = 3, "A&B" = 0.2))
residuals(fit)

}
\seealso{
\code{\link{plot.residuals.eulerr}}, \code{\link{eulerr}}
}

