% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glyph-grobs.R
\name{setup_euler_glyphs}
\alias{setup_euler_glyphs}
\title{Build a draw-time glyph layer}
\usage{
setup_euler_glyphs(
  x,
  glyphs,
  n_vertices,
  combo_labels,
  tags = NULL,
  set_labels = NULL,
  number
)
}
\description{
Build a draw-time glyph layer
}
\keyword{internal}
