% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glyph-grobs.R
\name{setup_euler_set_labels}
\alias{setup_euler_set_labels}
\title{Build draw-time exterior set labels}
\usage{
setup_euler_set_labels(x, labels, n_vertices, tags = NULL, number)
}
\description{
Build draw-time exterior set labels
}
\keyword{internal}
