% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_geometry.R
\name{setup_geometry}
\alias{setup_geometry}
\title{Compute geometries and label locations}
\usage{
setup_geometry(
  x,
  fills,
  edges,
  labels,
  quantities,
  annotations,
  n,
  merged_sets,
  placement_opts = NULL,
  do_complement_label = FALSE
)
}
\arguments{
\item{x}{an object of class 'euler'}

\item{fills}{fills}

\item{edges}{edges}

\item{labels}{labels}

\item{quantities}{quantities}

\item{annotations}{annotations}

\item{n}{number of vertices to use to render each ellipse}

\item{merged_sets}{which sets have been merged?}

\item{placement_opts}{label placement options, or \code{NULL} for the defaults}

\item{do_complement_label}{whether to place a label for the complement
region}
}
\value{
a list object with slots for the various objects
}
\description{
Compute geometries and label locations
}
\keyword{internal}
