% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{setup_gpar}
\alias{setup_gpar}
\title{Setup gpars}
\usage{
setup_gpar(x, n = 1, optional = list(), required = list())
}
\arguments{
\item{x}{input}
}
\value{
a list or FALSE
}
\description{
Setup gpars
}
\keyword{internal}
