% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layout-fixup.R
\name{shelf_pack}
\alias{shelf_pack}
\title{Shelf packing algorithm for packing rectangles in a bin.}
\usage{
shelf_pack(m)
}
\arguments{
\item{m}{A matrix of vertices for the rectangles.}
}
\value{
A list with new vertices and the new order of the rectangles.
}
\description{
Shelf packing algorithm for packing rectangles in a bin.
}
\keyword{internal}
