% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layout.R
\name{skyline_pack}
\alias{skyline_pack}
\title{Skyline packing algorithm}
\usage{
skyline_pack(m)
}
\arguments{
\item{m}{a matrix of vertices for the rectangles}
}
\value{
A matrix with updated vertices for the rectangles.
}
\description{
Skyline packing algorithm
}
\keyword{internal}
