% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_grobs.R
\name{stripe_segments}
\alias{stripe_segments}
\title{Grobify Euler objects}
\usage{
stripe_segments(xlim, ylim, angle = 45, spacing = NULL)
}
\arguments{
\item{x}{geometry data}

\item{fills}{fills params}

\item{patterns}{pattern params}

\item{edges}{edges params}

\item{labels}{labels params}

\item{quantities}{quantities params}

\item{number}{current diagram number}

\item{merged_sets}{sets that are the same and have been merged}
}
\value{
A \code{\link[grid:grid.grob]{grid::gList()}} is returned.
}
\description{
Grobify Euler objects
}
\keyword{internal}
