% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{update_list}
\alias{update_list}
\title{Update a List with User Input}
\usage{
update_list(x, val)
}
\arguments{
\item{x}{A list to be updated.}

\item{val}{Stuff to update \code{x} with.}
}
\value{
Returns an updated list.
}
\description{
Wrapper for \code{\link[utils:modifyList]{utils::modifyList()}}.
}
\seealso{
\code{\link[utils:modifyList]{utils::modifyList()}}
}
\keyword{internal}
