% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glyph-grobs.R
\name{validate_glyph_options}
\alias{validate_glyph_options}
\title{Validate glyph placement controls}
\usage{
validate_glyph_options(x)
}
\arguments{
\item{x}{normalized glyph options.}
}
\value{
\code{NULL}, invisibly.
}
\description{
Validate glyph placement controls
}
\keyword{internal}
