% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/euler-methods.R
\name{coef.euler}
\alias{coef.euler}
\title{Return ellipses from the euler object}
\usage{
\method{coef}{euler}(object, ...)
}
\arguments{
\item{object}{object of class \code{'euler'}}

\item{...}{ignored}
}
\value{
a data frame of the ellipses in the fit
}
\description{
Return ellipses from the euler object
}
\keyword{internal}
