% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layout-fixup.R
\name{compress_layout}
\alias{compress_layout}
\title{Compress a Euler Layout}
\usage{
compress_layout(fpar, id, fit)
}
\arguments{
\item{fpar}{A Euler layout fit with \code{\link[=euler]{euler()}}}

\item{id}{The binary index of sets.}
}
\value{
A modified fpar object.
}
\description{
Compress a Euler Layout
}
