% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{dont_print}
\alias{dont_print}
\title{Suppress Printing}
\usage{
dont_print(x, ...)
}
\arguments{
\item{x}{Object to (not) print.}

\item{...}{Arguments to \code{x}.}
}
\value{
Nothing, which is the point.
}
\description{
Suppress Printing
}
\keyword{internal}
