% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eulerr.R
\docType{package}
\name{eulerr-package}
\alias{eulerr}
\alias{eulerr-package}
\title{eulerr: Area-Proportional Euler and Venn Diagrams with Circles or Ellipses}
\description{
Generate area-proportional Euler diagrams
using numerical optimization. A Euler diagram is a generalization of a Venn
diagram, relaxing the criterion that all interactions need to be
represented. Diagrams may be fit with ellipses and circles via
a wide range of inputs and can be visualized in numerous ways.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/jolars/eulerr}
  \item Report bugs at \url{https://github.com/jolars/eulerr/issues}
}

}
\author{
\strong{Maintainer}: Johan Larsson \email{johanlarsson@outlook.com}

Other contributors:
\itemize{
  \item A. Jonathan R. Godfrey [contributor]
  \item Tim Kelley (Original Nelder-Mead code) [contributor]
  \item David H. Eberly (Geometric algorithms) [contributor]
}

}
\keyword{internal}
