% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eulerr.R
\docType{package}
\name{eulerr-package}
\alias{eulerr}
\alias{eulerr-package}
\title{eulerr: Area-Proportional Euler Diagrams with Circles or Ellipses}
\description{
If possible, generates exactly area-proportional Euler diagrams
using numerical optimization. A Euler diagram is a generalization of a Venn
diagram, relaxing the criterion that all interactions need to be
represented. The user if offered the choice between ellipses or circles.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/jolars/eulerr}
  \item Report bugs at \url{https://github.com/jolars/eulerr/issues}
}

}
\author{
\strong{Maintainer}: Johan Larsson \email{johanlarsson@outlook.com}

Other contributors:
\itemize{
  \item A. Jonathan R. Godfrey [contributor]
}

}
\keyword{internal}
