% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/euler.R
\name{eulerr}
\alias{eulerr}
\title{Area-Proportional Euler Diagrams (defunct)}
\usage{
eulerr(...)
}
\arguments{
\item{...}{Ignored}
}
\description{
\itemize{
\item Note: This function is defunct; please use \code{\link[=euler]{euler()}}.
instead.*
}
}
