% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eulerr_options.R
\name{eulerr_options}
\alias{eulerr_options}
\title{Get or set default graphical parameters for eulerr}
\usage{
eulerr_options(...)
}
\arguments{
\item{...}{various graphical parameters}
}
\value{
This function gets or sets updates in the global environment
that are used in \code{\link[=plot.euler]{plot.euler()}}.
}
\description{
Get or set default graphical parameters for eulerr
}
\examples{
eulerr_options(edges = list(col = "blue"), n = 300)
}
