% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tag-grobs.R
\name{find_eulertags}
\alias{find_eulertags}
\title{Find the \code{EulerTags} child of an \code{EulerPanel}, if any.}
\usage{
find_eulertags(panel)
}
\description{
Find the \code{EulerTags} child of an \code{EulerPanel}, if any.
}
\keyword{internal}
