% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/euler-methods.R
\name{fitted.euler}
\alias{fitted.euler}
\title{Fitted values of euler object}
\usage{
\method{fitted}{euler}(object, ...)
}
\arguments{
\item{object}{object of class \code{'euler'}}

\item{...}{ignored}
}
\value{
fitted values
}
\description{
Fitted values of euler object
}
\keyword{internal}
