% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geometry.R
\name{get_bounding_box}
\alias{get_bounding_box}
\title{Get the bounding box of an ellipse}
\usage{
get_bounding_box(h, k, a, b = NULL, phi = NULL)
}
\arguments{
\item{h}{x-coordinate for the center}

\item{k}{y-coordinate for the center}

\item{a}{radius or semi-major axis}

\item{b}{semi-minor axis}

\item{phi}{rotation}
}
\value{
The bounding box as a list with xlim and ylim
}
\description{
Get the bounding box of an ellipse
}
\keyword{internal}
