% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_integer}
\alias{is_integer}
\title{Check if a vector is an integer}
\usage{
is_integer(x, tol = .Machine$double.eps^0.5)
}
\arguments{
\item{x}{a vector}
}
\value{
TRUE of FALSE.
}
\description{
Check if a vector is an integer
}
\keyword{internal}
