% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tag-grobs.R
\name{makeContent.EulerTags}
\alias{makeContent.EulerTags}
\title{Avoid overlap for labels}
\usage{
\method{makeContent}{EulerTags}(x)
}
\description{
This method for \code{\link[grid:makeContent]{grid::makeContent()}} sets up
}
\keyword{internal}
