% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/label_placement.R
\name{measure_tag_sizes}
\alias{measure_tag_sizes}
\title{Measure all candidate tag sizes (regions + optional complement) inside
a fresh measurement viewport scaled to \code{xlim}/\code{ylim}.}
\usage{
measure_tag_sizes(
  centers,
  do_complement_label,
  complement_label,
  labels_gp,
  quantities_gp,
  padding,
  xlim,
  ylim
)
}
\description{
Measure all candidate tag sizes (regions + optional complement) inside
a fresh measurement viewport scaled to \code{xlim}/\code{ylim}.
}
\keyword{internal}
