% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{mix_colors}
\alias{mix_colors}
\title{Blend (Average) Colors for Use in Euler Diagrams}
\usage{
mix_colors(rcol_in)
}
\arguments{
\item{rcol_in}{A vector of R colors}
}
\value{
A single R color
}
\description{
Blend (Average) Colors for Use in Euler Diagrams
}
\keyword{internal}
