% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.euler.R
\name{panel.euler.circles}
\alias{panel.euler.circles}
\title{Panel function for Euler circles (deprecated)}
\usage{
panel.euler.circles(x, y, r, border = "black", fill = "transparent",
  fill_alpha = 0.4, ..., identifier = NULL, name.type = "panel", col,
  font, fontface)
}
\arguments{
\item{x}{X coordinates for the centers.}

\item{y}{Y coordinates for the centers.}

\item{r}{Radius of the circle}

\item{border}{Border color.}

\item{fill}{Circle fill.}

\item{fill_alpha}{Alpha (opacity) for the fill. (See \code{\link[grid:gpar]{grid::gpar()}}.)}

\item{...}{Passed on to \code{\link[grid:grid.circle]{grid::grid.circle()}}.}

\item{identifier}{A character string that is prepended to the name of the
grob that is created.}

\item{name.type}{A character value indicating whether the name of the grob
should have panel or strip information added to it. Typically either
\code{"panel"}, \code{"strip"}, \code{"strip.left"}, or \code{""} (for no extra information).}

\item{col}{Ignored}

\item{font}{Ignored}

\item{fontface}{Ignored}
}
\value{
Plots circles inside a trellis panel.
}
\description{
Panel function for Euler circles (deprecated)
}
\seealso{
\code{\link[grid:grid.circle]{grid::grid.circle()}}.
}
