% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.euler.R
\name{panel.euler.labels}
\alias{panel.euler.labels}
\title{Panel Function for Circle Labels}
\usage{
panel.euler.labels(x, y, ra, rb, phi, labels, counts = TRUE, original.values,
  fitted.values, ...)
}
\arguments{
\item{x}{X coordinates for the circle centers.}

\item{y}{Y coordinates for the circle centers.}

\item{ra}{Semi-major axes.}

\item{rb}{Semi-minor axes.}

\item{phi}{Rotation of the ellipse (as the counter-clockwise angle from
the positive x-axis to the semi-major axis).}

\item{labels}{Labels to plot on the circles.}

\item{counts}{Plots the original values for the disjoint set combinations
(\code{original.values}). Can also be a list, in which the contents of the list
will be passed on to \code{\link[lattice:panel.text]{lattice::panel.text()}} to modify the appearance of
the counts.}

\item{original.values}{Original values for the disjoint set combinations.}

\item{fitted.values}{Fitted values for the disjoint set combinations.}

\item{...}{Arguments passed on to \code{\link[=panel.text]{panel.text()}}}
}
\value{
Computes and plots labels or counts inside the centers of the
circles' overlaps.
}
\description{
Panel Function for Circle Labels
}
