% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot.euler}
\alias{plot.euler}
\title{Plot Euler diagrams}
\usage{
\method{plot}{euler}(x, fill = NULL, fill_opacity = 0.4, border = "black",
  lty = 1, lwd = 1, fontface = "plain", cex = 1, labels = NULL,
  key = FALSE, counts = FALSE, main = NULL, layout = NULL,
  outer_strips = TRUE, ...)
}
\arguments{
\item{x}{Euler diagram specification from \pkg{eulerr}}

\item{fill}{Colors to fill circles with.}

\item{fill_opacity}{Opacity of fill colors.}

\item{border}{Border color.}

\item{lty}{Line type(s) for circles as a integer or character vector. (See
\code{\link[grid]{gpar}}.)}

\item{lwd}{Line weight(s) for circles as a numeric vector. (See
\code{\link[grid]{gpar}}.)}

\item{fontface}{Fontface as a integer or character vector. (See
\code{\link[grid]{gpar}}.)}

\item{cex}{The relative size of text as a numeric vector. (See
\code{\link[grid]{gpar}}.)}

\item{labels}{Labels for the sets as a character vector.}

\item{key}{Set to \code{TRUE} to automatically generate a legend or input a
list to manually construct the legend (please see the entry
for \code{auto.key} in \code{\link[lattice]{xyplot}}). If \code{FALSE},
labels will be automatically generated inside the sets.}

\item{counts}{Set to \code{TRUE} to label set combinations with counts from
the input.}

\item{main}{Title of the plot.}

\item{layout}{Specifies the layout for the trellis panels if \code{by} was
used in the call to \code{\link{euler}}. This argument cannot be used
in conjunction with \code{outer_strips}, which will be forced to
\code{FALSE} if \code{layout} is not \code{NULL}.}

\item{outer_strips}{Set to \code{TRUE} to put second level factors on the
left margin of the trellis plot. (This argument is only considered if there
are 2 factors specified to \code{by} in the call to \code{\link{euler}}.)}

\item{\dots}{Arguments to pass to \code{\link[lattice]{xyplot}}.}
}
\value{
An object of class \code{trellis} from \pkg{lattice}, which has
\code{print} (plots the object and is called by default) and \code{update}
methods.
}
\description{
Plot Euler diagrams with trellis graphics from \pkg{lattice}. This function
calls \code{\link[lattice]{xyplot}} under the hood, allowing plots of
both individual euler diagrams as well as grids of diagrams
in case the \code{by} argument was used in the call to \code{\link{euler}}.
}
\details{
If \code{fill} is left blank, \pkg{eulerr} chooses color palettes based on
the number of sets and tries to provide palettes adapted to color vision
deficiencie based on functionality from \pkg{qualpalr}.

Several additional arguments can be handed down to
\code{\link[lattice]{xyplot}} via \ldots but be advised that this might cause
undesired conflicts with arguments that are provided inside this function.
It may be safer to use \code{\link[stats]{update}} to provide further
arguments instead.
}
\examples{
fit <- euler(c("A" = 10, "B" = 5, "A&B" = 3))
plot(fit, fill_opacity = 0.7)

# Change to italic roman font, remove borders and switch colors
plot(fit, fill = c("dodgerblue4", "darkgoldenrod1"), lwd = 0,
     fontface = "italic")

# Add counts to the plot
plot(fit, counts = TRUE)

# Add a custom legend and retain counts
plot(fit, counts = TRUE, key = list(space = "bottom", columns = 2))

# Plot without fills and distinguish sets with border types instead
plot(fit, lty = c("solid", "dotted"), fill = "transparent", cex = 2,
     fontface = 2, labels = c("foo", "bar"))

# Plot a grid of euler plots
dat <- data.frame(
  A      = sample(c(TRUE, FALSE), size = 100, replace = TRUE),
  B      = sample(c(TRUE, TRUE, FALSE), size = 100, replace = TRUE),
  gender = sample(c("Men", "Women"), size = 100, replace = TRUE),
  nation = sample(c("Sweden", "Denmark"), size = 100, replace = TRUE)
)

e_grid <- euler(dat[, 1:2], by = dat[, 3:4])
plot(e_grid, key = TRUE)

# We can modify the grid layout as well
plot(e_grid, layout = c(1, 4))

}
\seealso{
\code{\link[lattice]{xyplot}}, \code{\link[grid]{gpar}},
  \code{\link[grid]{grid.circle}}, \code{\link[lattice]{panel.xyplot}},
  \code{\link{euler}}, \code{\link[qualpalr]{qualpal}}
}
