% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot.euler}
\alias{plot.euler}
\title{euler plots}
\usage{
\method{plot}{euler}(x, fill_opacity = 0.4, polygon_args = list(),
  text_args = list(), mar = c(2, 2, 2, 2), counts = FALSE, main, mfrow,
  ...)
}
\arguments{
\item{x}{Specifications for a Euler diagram in the form of a object of class
\code{'euler'}}

\item{fill_opacity}{Opacity for the fill colors between 0 and 1. Values
beyond this range will be clamped to [0, 1] without warning.}

\item{polygon_args}{Arguments for \code{\link[graphics]{polygon}},
which is used to draw the circles.}

\item{text_args}{Arguments for \code{\link[graphics]{text}},
which is used to draw the text.}

\item{mar}{Margins for the plot area, set via\code{par()[["mar"]]}.}

\item{counts}{Plot counts for each unique section of the diagram. Thse are
the values from the original set specification.}

\item{main}{Title for the euler diagram(s).}

\item{mfrow}{Layout for the grid if a list of euler diagrams is provided.}

\item{\dots}{Arguments for \code{\link[graphics]{plot}} (that draws the plot
area).}
}
\description{
Plot Euler diagrams from euler using base R graphics.
}
\details{
If no color is specified to \code{polygon_args}, fills will be colored
using a predefined palette taken from \pkg{qualpal}.
}
\examples{
fit <- euler(c("A" = 10, "B" = 5, "A&B" = 3))
plot(fit, fill_opacity = 0.7)

# Change to italic roman font, remove borders and switch colors
plot(fit,
     polygon_args = list(col = c("dodgerblue4", "darkgoldenrod1"),
                         border = "transparent"),
     text_args = list(font = 3))

# Plot without fills and distinguish sets with border types instead
plot(fit,
     polygon_args = list(lty = c("solid", "dotted"), col = "transparent"),
     text_args = list(cex = 2, font = 2))

# Plot a grid of euler plots
dat <- data.frame(
  A      = sample(c(TRUE, FALSE), size = 100, replace = TRUE),
  B      = sample(c(TRUE, TRUE, FALSE), size = 100, replace = TRUE),
  gender = sample(c("Men", "Women"), size = 100, replace = TRUE),
  nation = sample(c("Sweden", "Denmark"), size = 100, replace = TRUE)
)

e_grid <- euler(dat[, 1:2], by = dat[, 3:4])
plot(e_grid)

# We can provide custom titles for our diagrams

plot(e_grid, main = c("A", "B", "C", "D"))

# and use any options that plot.euler takes

plot(e_grid, polygon_args = list(col = "transparent"))

# It is also possible to change grid layout

plot(e_grid, mfrow = c(1, 4))

}
\seealso{
\code{\link[graphics]{plot}}, \code{\link[graphics]{polygon}},
  \code{\link[graphics]{text}} \code{\link{euler}}
}

