% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.euler.R
\name{plot.eulergram}
\alias{plot.eulergram}
\title{Print (plot) Euler diagram}
\usage{
\method{plot}{eulergram}(x, newpage = TRUE, ...)
}
\arguments{
\item{x}{an object of class \code{'eulergram'}, usually the output of
\code{\link[=plot.euler]{plot.euler()}}.}

\item{...}{ignored}
}
\value{
A plot is drawn on the current device using \code{\link[grid:Grid]{grid::Grid()}} graphics
and a \code{\link[grid:gTree]{grid::gTree()}} object is returned invisibly.
}
\description{
This function is responsible for the actual drawing of
\code{'eulergram'} objects created through \code{\link[=plot.euler]{plot.euler()}}. \code{\link[=print.eulergram]{print.eulergram()}}
is simply an alias for \code{\link[=plot.eulergram]{plot.eulergram()}}.
}
\examples{
# save the plot as a gTree
grob <- print(plot(euler(c(A = 1))))

if (require(gridExtra)) {
  grid.arrange(grob, grob)
}
}
