% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot.eulerr}
\alias{plot.eulerr}
\title{Plot Euler diagrams}
\usage{
\method{plot}{eulerr}(eulerr, pal = "Accent", alpha = 80, ...)
}
\value{
Plots a eulerr diagram using Base R graphics.
}
\description{
Plot Euler diagrams
}
\examples{

}

