% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot.eulerr}
\alias{plot.eulerr}
\title{Plot Euler diagrams}
\usage{
\method{plot}{eulerr}(x, alpha = 0.4, ...)
}
\arguments{
\item{x}{euler diagram specifications of class \code{'eulerr'}}

\item{alpha}{Alpha for the colors to be plotted}

\item{...}{Arguments to pass to text, polygon, or plot.}
}
\value{
Plots euler diagram using Base R graphics.
}
\description{
Plot euler diagrams using base R graphics.
}
\examples{
fit <- eulerr(c(A = 10, B = 5, "A&B" = 3))
plot(fit, alpha = .7)

}

