% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot.residuals.eulerr}
\alias{plot.residuals.eulerr}
\title{Plot residuals from eulerr fit.}
\usage{
\method{plot}{residuals.eulerr}(x, ...)
}
\arguments{
\item{x}{Residuals from eulerr object produced by \code{residuals(eulerr(.))}}

\item{...}{Arguments passed to \code{link[graphics]{dotchart}}.}
}
\description{
Plot squared residuals from an object of class \code{eulerr} using a
cleveland dot plot via \code{link[graphics]{dotchart}}.
}
\examples{
fit <- eulerr(c(A = 1, B = 1, C = 1,
                "A&B" = 0.5, "A&C" = 0.5, "B&C" = 0.5,
                "A&B&C" = 0.25))
res <- resid(fit)
plot(res, pch = 19)#'

}

