% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.euler.R
\name{print.euler}
\alias{print.euler}
\title{Print Euler Fits}
\usage{
\method{print}{euler}(x, round = 3, ...)
}
\arguments{
\item{x}{Euler diagram specification from \code{\link[=euler]{euler()}}.}

\item{round}{Number of decimal places to round to.}

\item{...}{Arguments passed to \code{\link[base:print.data.frame]{base::print.data.frame()}}.}
}
\value{
Prints the results of the fit.
}
\description{
Prints a data frame of the original set relationships and the fitted
values as well as diagError and stress statistics.
}
