% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{regionError}
\alias{regionError}
\title{regionError}
\usage{
regionError(fit, orig)
}
\arguments{
\item{fit}{Fitted values}

\item{orig}{Original values}
}
\value{
regionError
}
\description{
regionError
}
\keyword{internal}
