% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{residuals.eulerr}
\alias{residuals.eulerr}
\title{Residuals from a eulerr fit}
\usage{
\method{residuals}{eulerr}(object, ...)
}
\arguments{
\item{object}{A eulerr object.}

\item{...}{Currently ignored.}
}
\value{
Residuals.
}
\description{
Residuals from a eulerr fit
}

