% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geometry.R
\name{separate_two_discs}
\alias{separate_two_discs}
\title{Optimize Distance Between Circles Based On Overlap}
\usage{
separate_two_discs(r1, r2, overlap)
}
\arguments{
\item{r1}{radius of circle one}

\item{r2}{radius of circle two}

\item{overlap}{overlap (area) of the two circles}
}
\value{
The necessary distance between sets to achieve the desired overlap
area.
}
\description{
Optimize Distance Between Circles Based On Overlap
}
\keyword{internal}
