% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_geometry.R
\name{setup_geometry}
\alias{setup_geometry}
\title{Compute geometries and label locations}
\usage{
setup_geometry(x, fills, edges, labels, quantities, n, id, merged_sets)
}
\arguments{
\item{x}{an object of class 'euler'}

\item{fills}{fills}

\item{edges}{edges}

\item{labels}{labels}

\item{quantities}{quantities}

\item{n}{number of sets}

\item{id}{identity matrix}

\item{merged_sets}{which sets have been merged?}
}
\value{
a list object with slots for the various objects
}
\description{
Compute geometries and label locations
}
\keyword{internal}
