% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.euler.R
\name{setup_geometry}
\alias{setup_geometry}
\title{Compute geometries and label locations}
\usage{
setup_geometry(x, fills, edges, labels, quantities, mode, n, id)
}
\arguments{
\item{x}{an object of class 'euler'}

\item{mode}{mode}

\item{do_fills}{do fills?}

\item{do_edges}{do edges?}

\item{do_labels}{do labels?}

\item{do_quantities}{do quantities?}
}
\value{
a list object with slots for the various objects
}
\description{
Compute geometries and label locations
}
\keyword{internal}
