% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{setup_gpar}
\alias{setup_gpar}
\title{Setup gpars}
\usage{
setup_gpar(default = list(), user = list(), n)
}
\arguments{
\item{default}{default values}

\item{user}{user-inputted values}

\item{n}{required number of items}
}
\value{
a \code{gpar} object
}
\description{
Setup gpars
}
\keyword{internal}
