% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.euler.R
\name{setup_grobs}
\alias{setup_grobs}
\title{Grobify Euler objects}
\usage{
setup_grobs(x, fills, edges, labels, quantities, number)
}
\arguments{
\item{x}{geometry data}

\item{fills}{fills params}

\item{edges}{edges params}

\item{labels}{labels params}

\item{quantities}{quantities params}

\item{number}{current diagram number}
}
\value{
A \code{\link[grid:gList]{grid::gList()}} is returned.
}
\description{
Grobify Euler objects
}
\keyword{internal}
