% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_grobs.R
\name{setup_grobs}
\alias{setup_grobs}
\title{Grobify Euler objects}
\usage{
setup_grobs(
  x,
  fills,
  edges,
  labels,
  quantities,
  number,
  adjust_labels,
  merged_sets
)
}
\arguments{
\item{x}{geometry data}

\item{fills}{fills params}

\item{edges}{edges params}

\item{labels}{labels params}

\item{quantities}{quantities params}

\item{number}{current diagram number}

\item{centers}{data.frame of data for labels and quantities}
}
\value{
A \code{\link[grid:grid.grob]{grid::gList()}} is returned.
}
\description{
Grobify Euler objects
}
\keyword{internal}
