% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.euler.R
\name{setup_grobs}
\alias{setup_grobs}
\title{Grobify Euler objects}
\usage{
setup_grobs(x, fills, edges, labels, quantities)
}
\arguments{
\item{x}{an \code{euler} object}
}
\value{
A \code{\link[grid:gList]{grid::gList()}} is returned.
}
\description{
Grobify Euler objects
}
\keyword{internal}
