% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tag-grobs.R
\name{setup_tag}
\alias{setup_tag}
\title{Setup grobs for labels (labels, quantities, percentages)}
\usage{
setup_tag(data, labels, quantities, number)
}
\arguments{
\item{data}{data for the locations of points and more}

\item{labels}{plot parameters for labels}

\item{quantities}{plot parameters for quantities}

\item{number}{panel number, used for naming the resulting grob}
}
\value{
A \code{\link[grid:gTree]{grid::gTree()}} object
}
\description{
Setup grobs for labels (labels, quantities, percentages)
}
\keyword{internal}
