% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{tally_combinations}
\alias{tally_combinations}
\title{Tally set relationships}
\usage{
tally_combinations(sets)
}
\arguments{
\item{sets}{A data.frame with set relationships and weights.}
}
\value{
Calls \code{\link[=euler]{euler()}} after the set relationships have been coerced to a
named numeric vector.
}
\description{
Tally set relationships
}
\keyword{internal}
