% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{update_list}
\alias{update_list}
\title{Update list with input}
\usage{
update_list(old, new)
}
\arguments{
\item{x}{a list to be updated}

\item{val}{stuff to update \code{x} with}
}
\value{
Returns an updated list.
}
\description{
A wrapper for \code{\link[utils:modifyList]{utils::modifyList()}} that attempts to coerce non-lists to
lists before updating.
}
\seealso{
\code{\link[utils:modifyList]{utils::modifyList()}}
}
\keyword{internal}
